refactor: move get-all-votes out of results dir
