+++ app/get-all-votes.ts
